# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "有来管理后台 API",
        "description": "基于 ThinkPHP 8 的 RBAC 权限管理系统后端 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "开发环境"
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "用户列表",
                "description": "分页查询用户列表",
                "operationId": "25bc25e670f79d21155e9bc5fd389851",
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "post": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "新增用户",
                "description": "创建用户",
                "operationId": "f7d7609d84d12f39d1289a0340c86217",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "获取用户详情",
                "description": "获取用户详情",
                "operationId": "bcd40e0c3ec14813f3fa0234a6389dab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "put": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "修改用户",
                "description": "更新用户",
                "operationId": "60342e26908dee117c0c76436bd9b052",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/users/{ids}": {
            "delete": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "删除用户",
                "description": "批量删除用户",
                "operationId": "bb86a474e198750cd3fdf272974ae7b7",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "用户ID，多个以英文逗号分割",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/users/{id}/password/reset": {
            "patch": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "重置密码",
                "description": "重置密码",
                "operationId": "d69a22101b3fc208fb4c9487ed9d30cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/users/{id}/status": {
            "patch": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "修改用户状态",
                "description": "修改状态",
                "operationId": "689683f5ca9dbebedd37ace8aa8919c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/api/v1/users/profile": {
            "get": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "获取个人中心用户信息",
                "description": "获取个人中心用户信息",
                "operationId": "e3065ccfe962cdec93df4d9aa9f5624e",
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "put": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "修改个人中心用户信息",
                "description": "个人中心修改用户信息",
                "operationId": "33b0daef4ef9640c503a21c07473f2de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/api/v1/users/{id}/form": {
            "get": {
                "tags": [
                    "02.用户接口"
                ],
                "summary": "获取用户表单数据",
                "description": "获取用户表单数据",
                "operationId": "eaf8eea8a1f7b30144f1f3ed0e84564c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        }
    },
...
}